Rename utils.airflow_utils to utils.airflow
- use DEFAULT_DATE in sql files dags
